Replace deprecated id with rawId (#589)
"id" has been deprecated in the Credential class and it was replaced by the new property "$rawId". This change was also included in several functional tests where Base64UrlSafe was being used previously. Now, 'rawId' is used throughout the system to maintain consistency. A validation to ensure the presence of a 'rawId' in the constructor of the Credential class has also been added.