Update serializers and remove PHPStan errors
This commit refines the Symfony Serializer in the ExtensionDescriptor class and upgrades several packages in the composer.json file. It also cleans up the phpstan-baseline.neon file by removing noise from a slew of PHPStan errors, which indicates an improvement in the code quality and robustness. This makes the PHPStan static analysis tool more effective and useful for future developments.